refactor(recommendation): 추천 페이지 전용 훅 위치 정리
